+++ PREFS/Test/test_read_create.py
@@ -8,8 +8,8 @@ sys.path.append(os.path.abspath(os.path.join('..')))
 import __init__ as PREFS
 
 ### Test single preferences and without cascade
-prefs = lambda: PREFS.read_prefs_file("prefs2")
-UserPrefs = PREFS.PREFS(prefs, filename="Prefs/prefs2", dictionary=False, interpret=True, verbose=False, cascade=True)#, filterPrefs=filterPrefs) # Change (dictionary, interpret, debug) to True to test it.
+prefs = lambda: PREFS.read_prefs_file("prefs2.prefs")
+UserPrefs = PREFS.PREFS(prefs, filename="Prefs/prefs2.prefs", interpret=True, verbose=False, cascade=True)#, filterPrefs=filterPrefs) # Change (dictionary, interpret, debug) to True to test it.
 
 
 def test_reading_overwrite():
@@ -19,14 +19,12 @@ def test_reading_overwrite():
 	assert UserPrefs.file == prefs() #Test file attribute
 
 def test_create_prefs():
-	txt = Path(f"{UserPrefs.filename}.{UserPrefs.extension}").read_text()
-	PREFSstr = PREFS.convert_to_prefs(prefs, first_line=True)
-	#print("txt: ")
-	#print("---------------")
-	#print("PREFSstr: ")
+	txt = Path(UserPrefs.filename).read_text()
+	PREFSstr = PREFS.convert_to_prefs(prefs)
+
+	#print(f"{txt=}\n-------\n{PREFSstr=}")
 
 	assert txt == PREFSstr
-
 
 if __name__ == "__main__":
 	test_reading_overwrite()